% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autuacoes}
\alias{autuacoes}
\title{Multas Ambientais Distribuidas por Bens tutelados.}
\format{
Uma tibble
\describe{
\item{dataAuto}{}
\item{dataPagamento}{}
\item{municipio}{}
\item{numAI}{}
\item{tipoInfracao}{}
\item{ultimaAtualizacaoRelatorio}{}
\item{uf}{}
\item{tipoAuto}{}
\item{moeda}{}
\item{situacaoDebito}{}
\item{enquadramentoLegal}{}
\item{nomeRazaoSocial}{}
\item{cpfCnpj}{}
\item{valorAuto}{}
\item{enquadramentoJuridico}{}
\item{code_muni}{}
}
}
\source{
IBAMA - \url{https://dados.gov.br/dataset/arrecadacao-de-multas-ambientais-bens-tutelados/resource/01e0e66d-0125-4997-b2b6-cbf29bb39302}
}
\usage{
autuacoes
}
\description{
Multas Ambientais Distribuidas por Bens tutelados.
}
\examples{
head(autuacoes)

}
\keyword{datasets}
