% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_id_municipios.R
\name{baixar_id_municipios}
\alias{baixar_id_municipios}
\title{Baixar codigos municipais do IBGE}
\usage{
baixar_id_municipios()
}
\value{
a data frame
}
\description{
Baixar codigos municipais do IBGE
}
