% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpar_dados.R
\name{clean_dataset}
\alias{clean_dataset}
\title{Limpeza das bases}
\usage{
clean_dataset(base)
}
\arguments{
\item{base}{Base de dados que deverá ser limpa}
}
\value{
a data frame
}
\description{
Limpeza das bases
}
\keyword{internal}
