% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{converter function's name}
\alias{converter function's name}
\alias{converter}
\title{Função de conversão para a moeda do Plano Real}
\usage{
converter(moeda, valor)
}
\arguments{
\item{moeda}{Moeda de base que será convertida}

\item{valor}{Quantia monetária.}
}
\value{
Retorna um valor único ou vetor convertido para Reais
}
\description{
Given the currency that was used, the function returns the values converted to Real (R$).
}
