% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\docType{data}
\name{df_arrec_convertido}
\alias{df_arrec_convertido}
\title{E se quiser adicionar como coluna da sua tibble, você pode fazer. Obs.: essa etapa demora pode demorar 5 minutos}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1275906 rows and 21 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1275906 rows and 21 columns.
}
\usage{
df_arrec_convertido

df_arrec_convertido
}
\description{
E se quiser adicionar como coluna da sua tibble, você pode fazer. Obs.: essa etapa demora pode demorar 5 minutos

É possivel converter multiplos valores simultaneamente?
}
\keyword{datasets}
