% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_multas_arrecadadas.R
\name{download_multas_arrecadadas}
\alias{download_multas_arrecadadas}
\title{Downloading data from Ibama database - Multas Arrecadadas}
\usage{
download_multas_arrecadadas()
}
\value{
a data frame
}
\description{
Downloading data from Ibama database - Multas Arrecadadas
}
