% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_multas_distribuidas.R
\name{download_multas_distribuidas}
\alias{download_multas_distribuidas}
\title{Downloading data from Ibama database - Multas Distribuidas}
\usage{
download_multas_distribuidas()
}
\value{
a data frame
}
\description{
Downloading data from Ibama database - Multas Distribuidas
}
