% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\docType{data}
\name{file_url}
\alias{file_url}
\title{Currency converter}
\format{
An object of class \code{character} of length 1.
}
\usage{
file_url
}
\arguments{
\item{moeda}{The .....  of currency.}

\item{valor}{The... monetary values.}
}
\value{
Retorna um valor único ou vetor convertido para Reais
Carregar dados
}
\description{
Given the currency that was used, the function returns the values converted to Real (R$).
Function was developed from the equations proposed at http://idealsoftwares.com.br/tabelas/tabela.php?id=351
}
\keyword{datasets}
