% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_data.R
\name{get_data_of_brazil}
\alias{get_data_of_brazil}
\title{Downloading data from Ibama database - Multas arrecadadas}
\usage{
get_data_of_brazil(estados = "all", tipo_multa)
}
\value{
A tibble
}
\description{
Downloading data from Ibama database - Multas arrecadadas
}
\keyword{internal}
