% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ibge_code}
\alias{ibge_code}
\title{Codigo do municipio segundo o IBGE}
\format{
Uma tibble, contendo:
\describe{
\item{code_muni}{Codigo do municipio segundo o IBGE}
\item{name_muni}{Nome do municipio}
\item{abrev_state}{Abreviacao do estado onde o municipio esta localizado}
}
}
\source{
Adaptado do pacote geobr - \url{https://github.com/ipeaGIT/geobr}
}
\usage{
ibge_code
}
\description{
Base auxiliar contendo o codigo do municipio
}
\examples{
head(ibge_code)
}
\keyword{datasets}
