% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juntar_bases.R
\name{juntar_bases}
\alias{juntar_bases}
\title{Realizar merging das bases}
\usage{
juntar_bases(base_ibama, base_ibgecode)
}
\arguments{
\item{base_ibama}{dados baixados do ibama}

\item{base_ibgecode}{dados baixados do ibge}
}
\value{
a data frame
}
\description{
Realizar merging das bases
}
