% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpar_dados.R
\name{limpar_dados}
\alias{limpar_dados}
\title{Limpeza das bases}
\usage{
limpar_dados(base)
}
\arguments{
\item{base}{Base de dados que deverá ser limpa}
}
\value{
a data frame
}
\description{
Limpeza das bases
}
\keyword{internal}
