% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\docType{data}
\name{multas}
\alias{multas}
\title{Currency converter}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 20 columns.
}
\usage{
multas
}
\arguments{
\item{moeda}{The .....  of currency.}

\item{valor}{The... monetary values.}
}
\value{
Retorna um valor único ou vetor convertido para Reais
Carregar dados de multas para exemplo
}
\description{
Given the currency that was used, the function returns the values converted to Real (R$).
}
\keyword{datasets}
