% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\docType{data}
\name{multas_convertido}
\alias{multas_convertido}
\title{Pode ainda obter os resultados via purrr. Obs.: essa etapa demora pode demorar 5 minutos}
\format{
An object of class \code{numeric} of length 1000.
}
\usage{
multas_convertido
}
\description{
Pode ainda obter os resultados via purrr. Obs.: essa etapa demora pode demorar 5 minutos
}
\keyword{datasets}
