% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_arquivos_arrec.R
\name{obter_arquivos_arrecadacao}
\alias{obter_arquivos_arrecadacao}
\title{Downloading data from Ibama database - Multas arrecadadas}
\usage{
obter_arquivos_arrecadacao()
}
\arguments{
\item{ufs}{bla bla bla}

\item{path}{Path where the jason files will be saved.}
}
\value{
a data frame
}
\description{
Downloading data from Ibama database - Multas arrecadadas
}
