% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_arquivos_autuacoes.R
\name{obter_arquivos_multas_distribuidas}
\alias{obter_arquivos_multas_distribuidas}
\title{Downloading data from Ibama database - multas distribuidas}
\usage{
obter_arquivos_multas_distribuidas()
}
\arguments{
\item{ufs}{bla bla bla}

\item{path}{Path where the jason files will be saved.}
}
\value{
a data frame
}
\description{
Downloading data from Ibama database - multas distribuidas
}
