% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_arquivos_arrecadacao.R
\name{obter_arrecadacao_estado}
\alias{obter_arrecadacao_estado}
\title{Obter dados de arrecadacao para um estado}
\usage{
obter_arrecadacao_estado(uf)
}
\arguments{
\item{uf}{Abreviation of the name of the state (two letters in lower case)}
}
\value{
A tibble
}
\description{
Obter dados de arrecadacao para um estado
}
\examples{
Ibamam:::obter_arrecadacao_estado("SP")
}
\keyword{internal}
