% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_arquivos_multas_distribuidas.R
\name{obter_multas_distribuidas_brasil}
\alias{obter_multas_distribuidas_brasil}
\title{Downloading data from Ibama database - multas distribuidas}
\usage{
obter_multas_distribuidas_brasil()
}
\value{
a data frame
}
\description{
Downloading data from Ibama database - multas distribuidas
}
\keyword{internal}
