% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_arquivos_multas_distribuidas.R
\name{obter_multas_distribuidas_estado}
\alias{obter_multas_distribuidas_estado}
\title{Obter dados de multas distribuidas para um estado}
\usage{
obter_multas_distribuidas_estado(uf)
}
\arguments{
\item{uf}{Abreviation of the name of the state (two letters in lower case)}
}
\value{
A tibble
}
\description{
Obter dados de multas distribuidas para um estado
}
\examples{
Ibamam:::obter_multas_distribuidas_estado("SP")
}
\keyword{internal}
