% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-datasets.R
\docType{data}
\name{uf_abrev}
\alias{uf_abrev}
\title{Vetor de siglas de estados brasileiros}
\format{
An object of class \code{character} of length 27.
}
\source{
Adaptado do pacote geobr - \url{https://github.com/ipeaGIT/geobr}
}
\usage{
uf_abrev
}
\description{
Vetor de siglas de estados brasileiros
}
\examples{
uf_abrev
}
\keyword{internal}
