% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-unite_with_ibge_code.R
\name{unite_with_ibge_code}
\alias{unite_with_ibge_code}
\title{Realizar merging das bases}
\usage{
unite_with_ibge_code(base_ibama, base_ibgecode)
}
\arguments{
\item{base_ibama}{dados baixados do ibama}

\item{base_ibgecode}{dados baixados do ibge}
}
\value{
a data frame
}
\description{
Realizar merging das bases
}
\keyword{internal}
