% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-get_dataset_ibamam.R
\name{get_dataset_ibamam}
\alias{get_dataset_ibamam}
\title{Downloading data from Ibama database - Multas Arrecadadas}
\usage{
get_dataset_ibamam(dataset, uf = "all")
}
\arguments{
\item{dataset}{Which dataset to search for. Can receive one of the values:
'arrecadadas' or 'distribuidas'}

\item{uf}{Abreviation of the name of the states (two letters in lower case)
that data will be obtained. Default is "all", so all data from all
the states of Brazil will be returned by default.}
}
\value{
a data frame
}
\description{
Downloading data from Ibama database - Multas Arrecadadas
}
