% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_multas_arrecadadas.R, R/converter.R
\docType{data}
\name{multas}
\alias{multas}
\title{Arrecadação de Multas Ambientais por Bens tutelados - Multas.}
\format{
\describe{
\item{dataAuto}{}
\item{dataPagamento}{}
\item{municipio}{}
\item{valorPago}{}
\item{numAI}{}
\item{tipoInfracao}{}
\item{ultimaAtualizacaoRelatorio}{}
\item{parcela}{}
\item{quantidadeParcela}{}
\item{uf}{}
\item{tipoAuto}{}
\item{moeda}{}
\item{statusDebito}{}
\item{enquadramentoLegal}{}
\item{nomeRazaoSocial}{}
\item{cpfCnpj}{}
\item{valorbaseParcela}{}
\item{valorAuto}{}
\item{enquadramentoJuridico}{}
\item{code_muni}{}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 20 columns.
}
\source{
\url{https://dados.gov.br/dataset/arrecadacao-de-multas-ambientais-bens-tutelados/resource/01e0e66d-0125-4997-b2b6-cbf29bb39302}
}
\usage{
multas

multas
}
\arguments{
\item{moeda}{The .....  of currency.}

\item{valor}{The... monetary values.}
}
\value{
Retorna um valor único ou vetor convertido para Reais
Carregar dados de multas para exemplo
}
\description{
Given the currency that was used, the function returns the values converted to Real (R$).
}
\keyword{datasets}
