% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\docType{data}
\name{multas}
\alias{multas}
\title{Carregar dados de multas para exemplo}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 20 columns.
}
\usage{
multas
}
\description{
Carregar dados de multas para exemplo
}
\keyword{datasets}
