% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\docType{data}
\name{multas_convertido}
\alias{multas_convertido}
\title{A função funciona para valores de comprimento 1.}
\format{
An object of class \code{numeric} of length 1000.
}
\usage{
multas_convertido
}
\description{
A função funciona para valores de comprimento 1.
}
\keyword{datasets}
