% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_arquivos_arrecadacao.R
\name{obter_arquivos_arrecadacao}
\alias{obter_arquivos_arrecadacao}
\title{Downloading data from Ibama database - Multas arrecadadas}
\usage{
obter_arquivos_arrecadacao()
}
\value{
A tibble
}
\description{
Downloading data from Ibama database - Multas arrecadadas
}
