% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_arquivos_multas_distribuidas.R
\name{obter_arquivos_multas_distribuidas}
\alias{obter_arquivos_multas_distribuidas}
\title{Downloading data from Ibama database - multas distribuidas}
\usage{
obter_arquivos_multas_distribuidas()
}
\value{
a data frame
}
\description{
Downloading data from Ibama database - multas distribuidas
}
